requirements.txt
Django==2.0.2
django-autoslug==1.9.3
django-autoslug-iplweb==1.9.4.dev0
django-tinymce==2.7.0
Pillow==5.0.0
pytz==2018.3
sorl-thumbnail==12.4.1
